requirements.txt
pytest==7.4.3
selenium==4.14.0
allure-pytest==2.9.45